+++ src/hooks/useSSE.ts
@@ -51,16 +51,23 @@ const useSSE = <T = any>(props: UseSSEProps<T>) => {
     try {
       setIsRequesting(true);
       setIsConnected(false);
+      let isError = false;
+      let response: Response | undefined = undefined;
       await fetchEventSource(connectUrl, {
         method: 'post',
         signal: abortCtrlRef.current.signal,
         openWhenHidden: true,
-        onopen: async (response: Response) => {
+        onopen: async (resp: Response) => {
+          response = resp;
+          if (!response.ok) {
+            isError = true;
+            return;
+          }
           setIsConnected(true);
           onopen?.(response);
         },
         onmessage(event) {
-          if (!isRequestingRef.current) {
+          if (isRequestingRef.current) {
             setIsRequesting(false);
           }
           try {
@@ -79,10 +86,10 @@ const useSSE = <T = any>(props: UseSSEProps<T>) => {
           }
         },
         onerror(error) {
-          setIsRequesting(false);
           onError?.(error);
         },
         onclose() {
+          if (isError) return;
           setIsRequesting(false);
           onClose?.();
         },
@@ -92,12 +99,18 @@ const useSSE = <T = any>(props: UseSSEProps<T>) => {
         },
         ...restOptions,
       });
-    } finally {
+      if (isError) {
+        throw response;
+      }
       setIsRequesting(false);
+    } catch (error) {
+      console.error('SSE connection error:', error);
+      setIsRequesting(false);
+      setIsConnected(false);
     }
   });
 
-  const close = useRefFunction(() => {
+  const abort = useRefFunction(() => {
     setIsConnected(false);
     if (isConnected && abortCtrlRef.current) {
       abortCtrlRef.current.abort();
@@ -108,13 +121,13 @@ const useSSE = <T = any>(props: UseSSEProps<T>) => {
     if (autoConnectRef.current) {
       connect();
     }
-    return close;
+    return abort;
   }, []);
 
   // 清理函数
   return {
     connect,
-    close,
+    abort,
     isRequesting,
     isConnected,
   };